refactoring de l'enfant, logique pour les regles de validation lors de l'ajout d'un enfant a un stage dans le panier + refactoring du glabelexceptionhandler pour envoyer l'erreur au front

+++ tfe-services/src/main/java/com/tfe/service/StageInstService.java
@@ -58,8 +58,14 @@ public class StageInstService {
     }
 
     public void deleteStageInstBtId(int id){
+        StageInstanceEntity entity = repository.findById(id)
+                .orElseThrow(() -> new StageNotFoundException(id));
+
         if(!repository.existsById(id)){
             throw new StageNotFoundException(id);
+        }
+        if(entity.getNbrInscrit() > 0){
+            throw new IllegalStateException("Impossible to delete an instance of stage if there is at least 1 subscription");
         }
         repository.deleteById(id);
     }
